refactor eagle infer tests: shared fixture + kits + overlap matrix
