+++ test/registered/spec/eagle/test_eagle_infer_a.py
@@ -1,202 +1,70 @@
-import random
+"""EAGLE3 spec-decoding: the overlap (spec v2) x no-overlap (spec v1) matrix.
+
+Both cases run the same full kit set on the same standard config (EAGLE3,
+topk=1, page_size=1); the only difference is ``disable_overlap``. Config-specific
+variants (other models / page sizes / backends / timeouts) live in
+test_eagle_infer_b.py.
+"""
+
 import unittest
 
-import sglang as sgl
 from sglang.srt.environ import envs
-from sglang.srt.utils.hf_transformers_utils import get_tokenizer
 from sglang.test.ci.ci_register import register_cuda_ci
-from sglang.test.test_utils import (
-    DEFAULT_DRAFT_MODEL_EAGLE,
-    DEFAULT_DRAFT_MODEL_EAGLE3,
-    DEFAULT_TARGET_MODEL_EAGLE,
-    DEFAULT_TARGET_MODEL_EAGLE3,
-    CustomTestCase,
+from sglang.test.kits.matched_stop_kit import MatchedStopMixin
+from sglang.test.kits.spec_server_kits import (
+    SpecAccuracyKit,
+    SpecCorrectnessKit,
+    SpecFeatureKit,
+    SpecLogprobKit,
+    SpecPenaltyKit,
+)
+from sglang.test.server_fixtures.spec_eagle_fixture import SpecEagleServerBase
+
+register_cuda_ci(est_time=900, stage="extra-a", runner_config="1-gpu-large")
+
+
+class _Eagle3Standard(SpecEagleServerBase):
+    """Standard EAGLE3 config shared by the overlap / no-overlap cases."""
+
+    spec_algo = "EAGLE3"
+    spec_steps = 5
+    spec_topk = 1
+    spec_tokens = 6
+    attention_backend = "flashinfer"
+    max_running_requests = 8
+    chunked_prefill_size = 1024
+    mem_fraction_static = 0.6  # leave room for the parity reference server
+
+    # Kit thresholds (EAGLE3 topk=1; tune against CI if needed).
+    acc_length_thres = 1.8
+    batch_accept_len_thres = 1.5
+    gsm8k_score_thres = 0.7
+    gsm8k_accept_len_thres = 1.8
+
+    # Busy-time pool accounting check (topk=1 only).
+    env_overrides = ((envs.SGLANG_ENABLE_STRICT_MEM_CHECK_DURING_BUSY, 1),)
+
+
+_KITS = (
+    SpecCorrectnessKit,
+    SpecAccuracyKit,
+    SpecLogprobKit,
+    SpecPenaltyKit,
+    SpecFeatureKit,
+    MatchedStopMixin,
 )
 
-register_cuda_ci(est_time=357, stage="extra-a", runner_config="1-gpu-large")
+
+class TestEagle3SpecOverlap(_Eagle3Standard, *_KITS):
+    """Spec v2 (overlap scheduler on)."""
+
+    disable_overlap = False
 
 
-class TestEAGLEEngine(CustomTestCase):
-    BASE_CONFIG = {
-        "model_path": DEFAULT_TARGET_MODEL_EAGLE,
-        "speculative_draft_model_path": DEFAULT_DRAFT_MODEL_EAGLE,
-        "speculative_algorithm": "EAGLE",
-        "speculative_num_steps": 5,
-        "speculative_eagle_topk": 4,
-        "speculative_num_draft_tokens": 8,
-        "mem_fraction_static": 0.7,
-        "cuda_graph_max_bs": 5,
-        "trust_remote_code": True,
-    }
-    NUM_CONFIGS = 2
+class TestEagle3SpecNoOverlap(_Eagle3Standard, *_KITS):
+    """Spec v1 (overlap scheduler off)."""
 
-    THRESHOLDS = {
-        "batch_avg_accept_len": 1.9,
-        "accept_len": 3.6,
-    }
-
-    @classmethod
-    def setUpClass(cls):
-        envs.SGLANG_ENABLE_SPEC_V2.set(False)
-
-    @classmethod
-    def tearDownClass(cls):
-        envs.SGLANG_ENABLE_SPEC_V2.clear()
-
-    def setUp(self):
-        self.prompt = "Today is a sunny day and I like"
-        self.sampling_params = {"temperature": 0, "max_new_tokens": 8}
-
-        ref_engine = sgl.Engine(
-            model_path=self.BASE_CONFIG["model_path"], cuda_graph_max_bs=1
-        )
-        self.ref_output = ref_engine.generate(self.prompt, self.sampling_params)["text"]
-        ref_engine.shutdown()
-
-    def test_correctness(self):
-        configs = [
-            # Basic config
-            self.BASE_CONFIG,
-            # Chunked prefill
-            {**self.BASE_CONFIG, "chunked_prefill_size": 4},
-        ]
-
-        for i, config in enumerate(configs[: self.NUM_CONFIGS]):
-            with self.subTest(i=i):
-                print(f"{config=}")
-                engine = sgl.Engine(**config, log_level="info", decode_log_interval=10)
-                try:
-                    self._test_single_generation(engine)
-                    self._test_first_token_finish(engine)
-                    self._test_batch_generation(engine)
-                    self._test_eos_token(engine)
-                    self._test_acc_length(engine)
-                finally:
-                    engine.flush_cache()  # check engine alive
-                    engine.shutdown()
-                print("=" * 100)
-
-    def _test_single_generation(self, engine):
-        output = engine.generate(self.prompt, self.sampling_params)["text"]
-        print(f"{output=}, {self.ref_output=}")
-        self.assertEqual(output, self.ref_output)
-
-    def _test_batch_generation(self, engine):
-        prompts = [
-            "Hello, my name is",
-            "The president of the United States is",
-            "The capital of France is",
-            "The future of AI is",
-        ]
-        params = {"temperature": 0, "max_new_tokens": 50}
-
-        outputs = engine.generate(prompts, params)
-        for prompt, output in zip(prompts, outputs):
-            print(f"Prompt: {prompt}")
-            print(f"Generated: {output['text']}")
-            print("-" * 40)
-
-        print(f"{engine.get_server_info()=}")
-
-        avg_spec_accept_length = engine.get_server_info()["internal_states"][0][
-            "avg_spec_accept_length"
-        ]
-        print(f"{avg_spec_accept_length=}")
-        self.assertGreater(
-            avg_spec_accept_length, self.THRESHOLDS["batch_avg_accept_len"]
-        )
-
-    def _test_first_token_finish(self, engine):
-        prompt = [
-            f"There are {i} apples on the table. How to divide them equally?"
-            for i in range(8)
-        ]
-        params = [
-            {"temperature": 0, "max_new_tokens": random.randint(1, 3)} for _ in range(8)
-        ]
-        outputs = engine.generate(prompt, params)
-        for i, output in enumerate(outputs):
-            print(f"Prompt: {prompt[i]}")
-            print(f"Generated: {output['text']}")
-            print("-" * 40)
-
-    def _test_eos_token(self, engine):
-        prompt = "[INST] <<SYS>>\nYou are a helpful assistant.\n<</SYS>>\nToday is a sunny day and I like [/INST]"
-        params = {
-            "temperature": 0.1,
-            "max_new_tokens": 1024,
-            "skip_special_tokens": False,
-        }
-
-        tokenizer = get_tokenizer(DEFAULT_TARGET_MODEL_EAGLE)
-        output = engine.generate(prompt, params)["text"]
-        print(f"{output=}")
-
-        tokens = tokenizer.encode(output, truncation=False)
-        self.assertNotIn(tokenizer.eos_token_id, tokens)
-
-    def _test_acc_length(self, engine):
-        prompt = [
-            "Human: Give me a fully functional FastAPI server. Show the python code.\n\nAssistant:",
-        ] * 5  # test batched generation
-        sampling_params = {"temperature": 0, "max_new_tokens": 512}
-        output = engine.generate(prompt, sampling_params)
-        output = output[0]
-
-        if "spec_verify_ct" in output["meta_info"]:
-            acc_length = (
-                output["meta_info"]["completion_tokens"]
-                / output["meta_info"]["spec_verify_ct"]
-            )
-        else:
-            acc_length = 1.0
-
-        speed = (
-            output["meta_info"]["completion_tokens"]
-            / output["meta_info"]["e2e_latency"]
-        )
-        print(f"{acc_length=:.4f}, {speed=}")
-
-        self.assertGreater(acc_length, self.THRESHOLDS["accept_len"])
-
-
-class TestEAGLEEngineTokenMap(TestEAGLEEngine):
-    BASE_CONFIG = {
-        "model_path": "meta-llama/Meta-Llama-3-8B-Instruct",
-        "speculative_draft_model_path": "lmsys/sglang-EAGLE-LLaMA3-Instruct-8B",
-        "speculative_algorithm": "EAGLE",
-        "speculative_num_steps": 5,
-        "speculative_eagle_topk": 4,
-        "speculative_num_draft_tokens": 8,
-        "speculative_token_map": "thunlp/LLaMA3-Instruct-8B-FR-Spec/freq_32768.pt",
-        "mem_fraction_static": 0.7,
-        "cuda_graph_max_bs": 5,
-        "dtype": "float16",
-    }
-    NUM_CONFIGS = 1
-    THRESHOLDS = {
-        "batch_avg_accept_len": 1.9,
-        "accept_len": 2.5,
-    }
-
-
-class TestEAGLE3Engine(TestEAGLEEngine):
-    BASE_CONFIG = {
-        "model_path": DEFAULT_TARGET_MODEL_EAGLE3,
-        "speculative_draft_model_path": DEFAULT_DRAFT_MODEL_EAGLE3,
-        "speculative_algorithm": "EAGLE3",
-        "speculative_num_steps": 5,
-        "speculative_eagle_topk": 16,
-        "speculative_num_draft_tokens": 64,
-        "mem_fraction_static": 0.7,
-        "cuda_graph_max_bs": 5,
-        "dtype": "float16",
-    }
-    NUM_CONFIGS = 1
-    THRESHOLDS = {
-        "batch_avg_accept_len": 1.75,
-        "accept_len": 3.1,
-    }
+    disable_overlap = True
 
 
 if __name__ == "__main__":
